Move CFF scaler implementation to skrifa
